fixed up a few things

+++ 2020Code/src/main/java/frc/robot/subsystems/DiscSpinner.java
@@ -8,14 +8,18 @@ import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
 
 public class DiscSpinner extends SubsystemBase{
     private WPI_TalonSRX spinner;
-    private final int THROTTLE = 1; 
+    private final double THROTTLE = 1; 
 
     public DiscSpinner(){
         spinner = new WPI_TalonSRX(Constants.SPINNER);
    }
 
-   public void start() {spinner.set(THROTTLE);}
+   public void start() {
+       spinner.set(THROTTLE);
+    }
 
-   public void stop() {spinner.set(0);}
+   public void stop() {
+       spinner.set(0);
+    }
 
 }
